Refact: commons loading
- Move sxmo icons in sxmo_icons.sh

This script should be loaded when icons are needed

- Move init env variable setups in a /etc/profile.d/sxmo_init.sh

This script is loaded by tinydm and for ssh/tty logins.

The logged in user can re-trigger check_sxmo_wm if they toggle wm.

- sxmo_commons.sh now only load aliases. It can be loaded by scripts to
ensure busybox compatibility

- moved some parts of ~/.config/sxmo/xinit to ~/.config/sxmo/profile

This file also is loaded just before starting sway. It goal is to setup
env variables dedicated to sxmo

~/.config/sxmo/xinit only goal is now to trigger some dwm dedicated
things (as ~/.config/sxmo/sway can start exec commands to)

To recap, loads orders are :

tinydm:
	/etc/profile.d/* # do not set SXMO_WM
	~/.profile

	sxmo_winit.sh:
		~/.config/sxmo/profile

	sxmo_xinit.sh:
		~/.config/sxmo/profile
		~/.config/sxmo/xinit # to start sxmo_hooks.sh

ssh/tty:
	/etc/profile.d/* # will set SXMO_WM
	~/.profile

- unify sxmo_winit.sh and sxmo_init.sh

+++ scripts/core/sxmo_keyboard.sh
@@ -7,17 +7,24 @@ SXMO_NO_ICONS=1 #just to make it a bit faster
 . "$(dirname "$0")/sxmo_common.sh"
 
 isopen() {
+	if [ -z "$KEYBOARD" ]; then
+		exit 0 # ssh/tty usage by example
+	fi
 	pidof "$KEYBOARD" > /dev/null
 }
 
 open() {
-	#Note: KEYBOARD_ARGS is not quoted by design as it may includes a pipe and further tools
-	# shellcheck disable=SC2086
-	isopen || eval "$KEYBOARD" $KEYBOARD_ARGS &
+	if [ -n "$KEYBOARD" ]; then
+		#Note: KEYBOARD_ARGS is not quoted by design as it may includes a pipe and further tools
+		# shellcheck disable=SC2086
+		isopen || eval "$KEYBOARD" $KEYBOARD_ARGS &
+	fi
 }
 
 close() {
-	pkill -f "$KEYBOARD"
+	if [ "$KEYBOARD" ]; then # avoid killing everything !
+		pkill -f "$KEYBOARD"
+	fi
 }
 
 if [ "$1" = "toggle" ]; then
